feat: new tool mp4ff-dump from example code
